requirements.txt
# Web Scraping
beautifulsoup4==4.12.3
requests==2.31.0
lxml==5.1.0

# Text Processing & Embeddings
sentence-transformers==3.3.1
transformers==4.45.2
torch==2.4.0
nltk==3.8.1
rank-bm25==0.2.2

# Image Processing & CLIP
Pillow==10.3.0
ftfy==6.2.0  # Required for CLIP
regex==2024.5.15  # Required for CLIP

# Vector Database
qdrant-client==1.12.1

# LLM Integration
litellm==1.55.8  # Unified API for LLM providers (Groq, OpenAI, etc.)
pydantic==2.10.6  # Fix for Gradio Gallery json_schema bug (GitHub issue #11084)

# Text Splitting
langchain-text-splitters==0.3.2

# UI
gradio==5.9.1

# Utilities
tqdm==4.66.4
numpy==1.26.4
pandas==2.2.3

# Development
pytest==8.3.4
black==24.10.0
